+++ src/routes/ConsortiumManager/settings/authorization-roles/UsersCapabilities/index.js
@@ -1,0 +1,2 @@
+export { UsersCapabilitiesCompare } from './UsersCapabilitiesCompare'
+export { UsersCapabilitiesCompareItems } from './UsersCapabilitiesCompareItems'